feat(EA-803): Update kubernetes action

+++ kubernetes/src/autoscale.js
@@ -1,8 +1,10 @@
 const exec = require('@actions/exec');
 const fs = require('fs');
 
-const removeAutoscale = async (deploymentName, deploymentType, permanentReplicas, dryRunArg) => {
+const getAutoscale = async (deploymentName) =>{
+  // We introduce this variable to get the output of the kubectl command.
   let errOutput = '';
+  // Trying to get autoscaler first. Only delete it if it exists.
   try {
     await exec.exec('kubectl', ['get', 'hpa', deploymentName, `--namespace=${deploymentName}`],
       {
@@ -14,26 +16,54 @@ const removeAutoscale = async (deploymentName, deploymentType, permanentReplicas
       });
   } catch (err) {
     if (errOutput.includes('(NotFound)')) {
-      return;
+      return false;
     }
     throw err;
   }
+  return true;
+}
 
+/**
+ * Only removes autoscale configuration if it exists.
+ * Manually scales the number of replicas to the provided value.
+ */
+const removeAutoscale = async (deploymentName, deploymentType, permanentReplicas, dryRunArg) => {
+  // Check if the autoscale configuration exists for deployment.
+  let autoscaleExists = await getAutoscale(deploymentName);
+  // If it doesn't exist there's nothing to do.
+  if(!autoscaleExists)
+  {
+    return;
+  }
+  // Delete existing autoscaler configuration.
   await exec.exec('kubectl', ['delete', 'hpa', deploymentName, `--namespace=${deploymentName}`, ...dryRunArg]);
+  // Manually scale replicas to the provided value.
   await exec.exec('kubectl', ['scale', deploymentType, deploymentName, `--namespace=${deploymentName}`, `--replicas=${permanentReplicas}`, ...dryRunArg]);
 };
 
-
+/**
+ * Applies autoscale configuration to the existing deployment. 
+ * If autoscale configuration is not provided, autoscaling will be deleted from services' spec.
+ * Provided deploymentType must be consistent with the one of the existing service.
+ * 
+ * @param  deploymentName service name
+ * @param  deploymentType statefulSet or deployment
+ * @param  autoscale configuration must include minReplicas, maxReplicas and cpuPercent
+ * @param  permanentReplicas value to be used if autoscale configuration is deleted
+ * @param  dryRun if true no actual changes will be applied to production
+ */
 const applyAutoscale = async (deploymentName, deploymentType, autoscale, permanentReplicas,
   dryRun) => {
   const dryRunArg = dryRun ? ['--dry-run=client'] : [];
 
+  // If there's no autoscale parameters provided we remove autoscale configuration from deployment
   if (autoscale == null) {
     await removeAutoscale(deploymentName, deploymentType, permanentReplicas, dryRunArg);
     return;
   }
 
-  const hpaYaml = `
+//Horizontal pod autoscaler spec template.
+const hpaYaml = `
 apiVersion: autoscaling/v1
 kind: HorizontalPodAutoscaler
 metadata:
@@ -55,8 +85,10 @@ spec:
         averageUtilization: ${autoscale.cpuPercent}
 `;
 
+  // Write hpa template to disc to apply afterwards.
   fs.writeFileSync('hpa.yml', hpaYaml);
 
+  // Apply the autoscale configuration from above.
   await exec.exec('kubectl', ['apply', '-f', 'hpa.yml', ...dryRunArg]);
 };
